Move TimeSnapshot into statime-base
