+++ ntp-proto/src/system.rs
@@ -1,5 +1,5 @@
 use serde::{Deserialize, Serialize};
-use statime_base::{Duration, LeapStatus, TAI, Timestamp};
+use statime_base::{LeapStatus, TimeSnapshot};
 use std::collections::HashMap;
 use std::fmt::Debug;
 use std::net::{IpAddr, SocketAddr};
@@ -13,61 +13,6 @@ use crate::{
     identifiers::ReferenceId,
     source::{NtpSource, NtpSourceActionIterator, ProtocolVersion, SourceNtsData},
 };
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq)]
-pub struct TimeSnapshot {
-    /// Precision of the local clock
-    pub precision: Duration,
-    /// Current root delay
-    pub root_delay: Duration,
-    /// t=0 for root variance calculation
-    pub root_variance_base_time: Timestamp<TAI>,
-    /// Constant contribution for root variance
-    pub root_variance_base: f64,
-    /// Linear (*t) contribution for root variance
-    pub root_variance_linear: f64,
-    /// Quadratic (*t*t) contribution for root variance
-    pub root_variance_quadratic: f64,
-    /// Cubic (*t*t*t) contribution for root variance
-    pub root_variance_cubic: f64,
-    /// Current leap indicator state
-    pub leap_indicator: Option<LeapStatus>,
-    /// Total amount that the clock has stepped
-    pub accumulated_steps: Duration,
-    /// Crossing this amount of stepping will cause a Panic
-    pub accumulated_steps_threshold: Option<Duration>,
-}
-
-impl TimeSnapshot {
-    pub fn root_dispersion(&self, now: Timestamp<TAI>) -> Duration {
-        let t = (now - self.root_variance_base_time).as_seconds();
-        // Note: dispersion is the standard deviation, so we need a sqrt here.
-        Duration::from_f64_seconds(
-            (self.root_variance_base
-                + t * self.root_variance_linear
-                + t.powi(2) * self.root_variance_quadratic
-                + t.powi(3) * self.root_variance_cubic)
-                .sqrt(),
-        )
-    }
-}
-
-impl Default for TimeSnapshot {
-    fn default() -> Self {
-        Self {
-            precision: Duration::from_seconds_nanos(0, 1),
-            root_delay: Duration::ZERO,
-            root_variance_base_time: Timestamp::UNIX_EPOCH,
-            root_variance_base: 0.0,
-            root_variance_linear: 0.0,
-            root_variance_quadratic: 0.0,
-            root_variance_cubic: 0.0,
-            leap_indicator: None,
-            accumulated_steps: Duration::ZERO,
-            accumulated_steps_threshold: None,
-        }
-    }
-}
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, Default)]
 pub struct SystemSnapshot {
